requirements.txt
fastapi
uvicorn
python-dotenv
moviepy==1.0.3
minio
requests
python-multipart
numpy
decorator
imageio
imageio-ffmpeg
proglog
tqdm 